Began Refactoring
