+++ Unity-Wiimote.sublime-project
@@ -1,0 +1,10 @@
+{
+	"folders":
+	[
+		{
+			"follow_symlinks": true,
+			"path": "."
+		}
+	],
+	"solution_file": "./UnityVS.Unity-Wiimote.sln"
+}
